Refactor database module structure and update imports
- Moved database-related functions and classes into a new `database.py` module within the `utils/database` directory for better organization.
- Updated import paths across various modules to reflect the new structure, ensuring consistent access to database functions.
- Introduced a new `pallet_data.py` module to define enums and classes related to pallet management, enhancing code clarity and maintainability.
- Adjusted the import helper to accommodate the new database structure and improve code completion suggestions.
