+++ main.py
@@ -34,7 +34,7 @@ from utils import global_vars
 from utils.app_control import init_settings
 from utils.logging_config import setup_logger
 from utils.app_initialization import parse_arguments, initialize_app, setup_initial_app_state
-from utils.database import create_database
+from utils.database.database import create_database
 from utils.robot_control import update_database_from_usb
 from utils.ui_setup import (initialize_main_window, setup_input_validation, connect_signal_handlers,
                           setup_password_handling, setup_components, start_background_tasks,
