+++ utils/robot_control.py
@@ -2,7 +2,7 @@ import socket
 import logging
 import os
 from utils import global_vars
-from utils.database import save_to_database, find_file_in_database, list_available_files
+from utils.database.database import save_to_database, find_file_in_database, list_available_files
 from utils.status_manager import update_status_label
 
 logger = logging.getLogger(__name__)
